+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.5'
 }
 
-group = 'com.test'
+group = 'io.swaggy.swagger'
 version = '0.0.1-SNAPSHOT'
 
 java {
