+++ src/main/java/io/swaggy/swagger/customlib/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.test.swaggercustom.config;
+package io.swaggy.swagger.customlib.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
